testing/20151117.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20151117</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<h1 id="w3c/testing.20151117.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2015-11-17 / <a href="#end">end</a></h1>
<p>Options: 
<p><a href=20151116>Previous day</a>, <a href=20151118>Next day</a></p>
<ol id=lines>
<li id=l-1 class=hide><a href=#l-1>#</a> Session Start: Tue Nov 17 00:00:00 2015</li>
<li id=l-2 class=hide><a href=#l-2>#</a> Session Ident: #testing</li>
<li id=l-3 class=hide><a href=#l-3>#</a> [00:12] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-4 class=hide><a href=#l-4>#</a> [00:19] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-5 class=hide><a href=#l-5>#</a> [00:34] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-6 class=hide><a href=#l-6>#</a> [00:49] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-7 class=hide><a href=#l-7>#</a> [01:02] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-8 class=hide><a href=#l-8>#</a> [01:49] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-9 class=hide><a href=#l-9>#</a> [02:45] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-10 class=hide><a href=#l-10>#</a> [04:35] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-11 class=hide><a href=#l-11>#</a> [04:59] * Joins: fwtnb_ (~fwtnb@public.cloak)</li>
<li id=l-12 class=hide><a href=#l-12>#</a> [04:59] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-13 class=hide><a href=#l-13>#</a> [05:39] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-14 class=hide><a href=#l-14>#</a> [06:10] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-15 class=hide><a href=#l-15>#</a> [06:10] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-16 class=hide><a href=#l-16>#</a> [07:17] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-17 class=hide><a href=#l-17>#</a> [07:17] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-18 class=hide><a href=#l-18>#</a> [07:32] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-19 class=hide><a href=#l-19>#</a> [07:37] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-20 class=hide><a href=#l-20>#</a> [08:33] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-21 class=hide><a href=#l-21>#</a> [08:42] * Quits: rniwa (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-22 class=hide><a href=#l-22>#</a> [09:15] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-23 class=hide><a href=#l-23>#</a> [09:43] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-24 class=hide><a href=#l-24>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-25><a href=#l-25>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #2335: Implement 'labels' attribute on 'labelable elements' (master...sync_c2bf48e0d59a2a9e0d404f0530218baaa3b3c122) <a href="https://github.com/w3c/web-platform-tests/pull/2335">https://github.com/w3c/web-platform-tests/pull/2335</a></li>
<li id=l-26 class=hide><a href=#l-26>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-27 class=hide><a href=#l-27>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-28><a href=#l-28>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2335: Implement 'labels' attribute on 'labelable elements' (master...sync_c2bf48e0d59a2a9e0d404f0530218baaa3b3c122) <a href="https://github.com/w3c/web-platform-tests/pull/2335">https://github.com/w3c/web-platform-tests/pull/2335</a></li>
<li id=l-29 class=hide><a href=#l-29>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-30 class=hide><a href=#l-30>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-31><a href=#l-31>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_c2bf48e0d59a2a9e0d404f0530218baaa3b3c122 at c2bf48e: <a href="https://github.com/w3c/web-platform-tests/commit/c2bf48e">https://github.com/w3c/web-platform-tests/commit/c2bf48e</a></li>
<li id=l-32 class=hide><a href=#l-32>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-33 class=hide><a href=#l-33>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-34><a href=#l-34>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_7d4d65a243623b6b0b6da4a4d90ea2a0ae43a1f9 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/7d4d65a24362">https://github.com/w3c/web-platform-tests/commit/7d4d65a24362</a></li>
<li id=l-35><a href=#l-35>#</a> [09:54] -gitbot:#testing- web-platform-tests/sync_7d4d65a243623b6b0b6da4a4d90ea2a0ae43a1f9 7d4d65a Eli Friedman: Use attribute getter/setter macros for misc DOM attributes....</li>
<li id=l-36 class=hide><a href=#l-36>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-37 class=hide><a href=#l-37>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-38><a href=#l-38>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2336: Use attribute getter/setter macros for misc DOM attributes. (master...sync_7d4d65a243623b6b0b6da4a4d90ea2a0ae43a1f9) <a href="https://github.com/w3c/web-platform-tests/pull/2336">https://github.com/w3c/web-platform-tests/pull/2336</a></li>
<li id=l-39 class=hide><a href=#l-39>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-40 class=hide><a href=#l-40>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-41><a href=#l-41>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_7d4d65a243623b6b0b6da4a4d90ea2a0ae43a1f9 at 7d4d65a: <a href="https://github.com/w3c/web-platform-tests/commit/7d4d65a">https://github.com/w3c/web-platform-tests/commit/7d4d65a</a></li>
<li id=l-42 class=hide><a href=#l-42>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-43 class=hide><a href=#l-43>#</a> [09:54] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-44><a href=#l-44>#</a> [09:54] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_71075ab706f68bc24ea805ccd2de4fcafded597d (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/71075ab706f6">https://github.com/w3c/web-platform-tests/commit/71075ab706f6</a></li>
<li id=l-45><a href=#l-45>#</a> [09:54] -gitbot:#testing- web-platform-tests/sync_71075ab706f68bc24ea805ccd2de4fcafded597d 71075ab Nikki: Fixes #8213: Implement Blob variant of WebSocket.send()</li>
<li id=l-46 class=hide><a href=#l-46>#</a> [09:54] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-47 class=hide><a href=#l-47>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-48><a href=#l-48>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2337: Fixes #8213: Implement Blob variant of WebSocket.send() (master...sync_71075ab706f68bc24ea805ccd2de4fcafded597d) <a href="https://github.com/w3c/web-platform-tests/pull/2337">https://github.com/w3c/web-platform-tests/pull/2337</a></li>
<li id=l-49 class=hide><a href=#l-49>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-50 class=hide><a href=#l-50>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-51><a href=#l-51>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #2338: wpt for javascript urls with queries and fragments (master...sync_a08425822b21ff2a48c0752c0f839b347cc5eb43) <a href="https://github.com/w3c/web-platform-tests/pull/2338">https://github.com/w3c/web-platform-tests/pull/2338</a></li>
<li id=l-52 class=hide><a href=#l-52>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-53 class=hide><a href=#l-53>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-54><a href=#l-54>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_a08425822b21ff2a48c0752c0f839b347cc5eb43 at a084258 (+0 new commits): <a href="https://github.com/w3c/web-platform-tests/commits/sync_a08425822b21ff2a48c0752c0f839b347cc5eb43">https://github.com/w3c/web-platform-tests/commits/sync_a08425822b21ff2a48c0752c0f839b347cc5eb43</a></li>
<li id=l-55 class=hide><a href=#l-55>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-56 class=hide><a href=#l-56>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-57><a href=#l-57>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_5b042b7d1bf6df5695ffa322a87ceb7f4120fb23 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/5b042b7d1bf6">https://github.com/w3c/web-platform-tests/commit/5b042b7d1bf6</a></li>
<li id=l-58><a href=#l-58>#</a> [09:55] -gitbot:#testing- web-platform-tests/sync_5b042b7d1bf6df5695ffa322a87ceb7f4120fb23 5b042b7 Corey Farwell: Implement &lt;hr&gt; 'color' attribute</li>
<li id=l-59 class=hide><a href=#l-59>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-60 class=hide><a href=#l-60>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-61><a href=#l-61>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2339: Implement &lt;hr&gt; 'color' attribute (master...sync_5b042b7d1bf6df5695ffa322a87ceb7f4120fb23) <a href="https://github.com/w3c/web-platform-tests/pull/2339">https://github.com/w3c/web-platform-tests/pull/2339</a></li>
<li id=l-62 class=hide><a href=#l-62>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-63 class=hide><a href=#l-63>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-64><a href=#l-64>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_5b042b7d1bf6df5695ffa322a87ceb7f4120fb23 at 5b042b7: <a href="https://github.com/w3c/web-platform-tests/commit/5b042b7">https://github.com/w3c/web-platform-tests/commit/5b042b7</a></li>
<li id=l-65 class=hide><a href=#l-65>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-66 class=hide><a href=#l-66>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-67><a href=#l-67>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_8ee5fef1ec2ca982ce80cd66eedc523a0d0a77d2 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/8ee5fef1ec2c">https://github.com/w3c/web-platform-tests/commit/8ee5fef1ec2c</a></li>
<li id=l-68><a href=#l-68>#</a> [09:55] -gitbot:#testing- web-platform-tests/sync_8ee5fef1ec2ca982ce80cd66eedc523a0d0a77d2 8ee5fef Josh Matthews: Add test for OnErrorEventHandler special case.</li>
<li id=l-69 class=hide><a href=#l-69>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-70 class=hide><a href=#l-70>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-71><a href=#l-71>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2340: Add test for OnErrorEventHandler special case. (master...sync_8ee5fef1ec2ca982ce80cd66eedc523a0d0a77d2) <a href="https://github.com/w3c/web-platform-tests/pull/2340">https://github.com/w3c/web-platform-tests/pull/2340</a></li>
<li id=l-72 class=hide><a href=#l-72>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-73 class=hide><a href=#l-73>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-74><a href=#l-74>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/af9d938ba720259adbbaf3a1689614eb9c814889">https://github.com/w3c/web-platform-tests/commit/af9d938ba720259adbbaf3a1689614eb9c814889</a></li>
<li id=l-75><a href=#l-75>#</a> [09:55] -gitbot:#testing- web-platform-tests/master af9d938 Ms2ger: Merge pull request #2340 from w3c/sync_8ee5fef1ec2ca982ce80cd66eedc523a0d0a77d2...</li>
<li id=l-76 class=hide><a href=#l-76>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-77 class=hide><a href=#l-77>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-78><a href=#l-78>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_c878c22218b9ab30b828b1dde7c380df8a6f115e (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/c878c22218b9">https://github.com/w3c/web-platform-tests/commit/c878c22218b9</a></li>
<li id=l-79><a href=#l-79>#</a> [09:55] -gitbot:#testing- web-platform-tests/sync_c878c22218b9ab30b828b1dde7c380df8a6f115e c878c22 Manish Goregaokar: Add tests for get/set/removeAttributeNode()...</li>
<li id=l-80 class=hide><a href=#l-80>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-81 class=hide><a href=#l-81>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-82><a href=#l-82>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2341: Add tests for get/set/removeAttributeNode() (master...sync_c878c22218b9ab30b828b1dde7c380df8a6f115e) <a href="https://github.com/w3c/web-platform-tests/pull/2341">https://github.com/w3c/web-platform-tests/pull/2341</a></li>
<li id=l-83 class=hide><a href=#l-83>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-84 class=hide><a href=#l-84>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-85><a href=#l-85>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_c878c22218b9ab30b828b1dde7c380df8a6f115e at c878c22: <a href="https://github.com/w3c/web-platform-tests/commit/c878c22">https://github.com/w3c/web-platform-tests/commit/c878c22</a></li>
<li id=l-86 class=hide><a href=#l-86>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-87 class=hide><a href=#l-87>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-88><a href=#l-88>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_09171b14cb0e5089f9d80640194ccc9a37be684a (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/09171b14cb0e">https://github.com/w3c/web-platform-tests/commit/09171b14cb0e</a></li>
<li id=l-89><a href=#l-89>#</a> [09:55] -gitbot:#testing- web-platform-tests/sync_09171b14cb0e5089f9d80640194ccc9a37be684a 09171b1 Corey Farwell: Implement &lt;hr&gt; element 'width' attribute</li>
<li id=l-90 class=hide><a href=#l-90>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-91 class=hide><a href=#l-91>#</a> [09:55] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-92><a href=#l-92>#</a> [09:55] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2342: Implement &lt;hr&gt; element 'width' attribute (master...sync_09171b14cb0e5089f9d80640194ccc9a37be684a) <a href="https://github.com/w3c/web-platform-tests/pull/2342">https://github.com/w3c/web-platform-tests/pull/2342</a></li>
<li id=l-93 class=hide><a href=#l-93>#</a> [09:55] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-94 class=hide><a href=#l-94>#</a> [09:56] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-95><a href=#l-95>#</a> [09:56] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_09171b14cb0e5089f9d80640194ccc9a37be684a at 09171b1: <a href="https://github.com/w3c/web-platform-tests/commit/09171b1">https://github.com/w3c/web-platform-tests/commit/09171b1</a></li>
<li id=l-96 class=hide><a href=#l-96>#</a> [09:56] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-97 class=hide><a href=#l-97>#</a> [09:56] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-98><a href=#l-98>#</a> [09:56] -gitbot:#testing- [web-platform-tests] Ms2ger created sync_647963430dac82928c49ce9853e1bf1e26ddfa1c (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/647963430dac">https://github.com/w3c/web-platform-tests/commit/647963430dac</a></li>
<li id=l-99><a href=#l-99>#</a> [09:56] -gitbot:#testing- web-platform-tests/sync_647963430dac82928c49ce9853e1bf1e26ddfa1c 6479634 Corey Farwell: Make util::str::parse_length work with floats</li>
<li id=l-100 class=hide><a href=#l-100>#</a> [09:56] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-101 class=hide><a href=#l-101>#</a> [09:56] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-102><a href=#l-102>#</a> [09:56] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2343: Make util::str::parse_length work with floats (master...sync_647963430dac82928c49ce9853e1bf1e26ddfa1c) <a href="https://github.com/w3c/web-platform-tests/pull/2343">https://github.com/w3c/web-platform-tests/pull/2343</a></li>
<li id=l-103 class=hide><a href=#l-103>#</a> [09:56] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-104 class=hide><a href=#l-104>#</a> [09:56] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-105><a href=#l-105>#</a> [09:56] -gitbot:#testing- [web-platform-tests] Ms2ger deleted sync_647963430dac82928c49ce9853e1bf1e26ddfa1c at 6479634: <a href="https://github.com/w3c/web-platform-tests/commit/6479634">https://github.com/w3c/web-platform-tests/commit/6479634</a></li>
<li id=l-106 class=hide><a href=#l-106>#</a> [09:56] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-107 class=hide><a href=#l-107>#</a> [10:05] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-108><a href=#l-108>#</a> [10:05] -gitbot:#testing- [web-platform-tests] annevk deleted urlsearchparams at da89213: <a href="https://github.com/w3c/web-platform-tests/commit/da89213">https://github.com/w3c/web-platform-tests/commit/da89213</a></li>
<li id=l-109 class=hide><a href=#l-109>#</a> [10:05] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-110 class=hide><a href=#l-110>#</a> [10:11] * Joins: dom (dom@public.cloak)</li>
<li id=l-111 class=hide><a href=#l-111>#</a> [11:27] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-112><a href=#l-112>#</a> [11:27] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #1278: Make new Blob throw (master...non-new-Blob) <a href="https://github.com/w3c/web-platform-tests/pull/1278">https://github.com/w3c/web-platform-tests/pull/1278</a></li>
<li id=l-113 class=hide><a href=#l-113>#</a> [11:27] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-114 class=hide><a href=#l-114>#</a> [11:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-115><a href=#l-115>#</a> [11:29] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2334: localStorage and document.domain test (master...storage-document-domain) <a href="https://github.com/w3c/web-platform-tests/pull/2334">https://github.com/w3c/web-platform-tests/pull/2334</a></li>
<li id=l-116 class=hide><a href=#l-116>#</a> [11:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-117 class=hide><a href=#l-117>#</a> [11:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-118><a href=#l-118>#</a> [11:29] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/ca1843553b87a2198df35a7ebe3371284ed2ecf5">https://github.com/w3c/web-platform-tests/commit/ca1843553b87a2198df35a7ebe3371284ed2ecf5</a></li>
<li id=l-119><a href=#l-119>#</a> [11:29] -gitbot:#testing- web-platform-tests/master ca18435 Ms2ger: Merge pull request #2334 from w3c/storage-document-domain...</li>
<li id=l-120 class=hide><a href=#l-120>#</a> [11:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-121 class=hide><a href=#l-121>#</a> [11:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-122><a href=#l-122>#</a> [11:29] -gitbot:#testing- [web-platform-tests] Ms2ger deleted storage-document-domain at ac4fb38: <a href="https://github.com/w3c/web-platform-tests/commit/ac4fb38">https://github.com/w3c/web-platform-tests/commit/ac4fb38</a></li>
<li id=l-123 class=hide><a href=#l-123>#</a> [11:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-124 class=hide><a href=#l-124>#</a> [11:30] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-125><a href=#l-125>#</a> [11:30] -gitbot:#testing- [web-platform-tests] annevk deleted non-new-Blob at d8d1e56: <a href="https://github.com/w3c/web-platform-tests/commit/d8d1e56">https://github.com/w3c/web-platform-tests/commit/d8d1e56</a></li>
<li id=l-126 class=hide><a href=#l-126>#</a> [11:30] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-127 class=hide><a href=#l-127>#</a> [12:01] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-128><a href=#l-128>#</a> [12:01] -gitbot:#testing- [web-platform-tests] annevk created noreferrer-name-reuse (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/8ea6709d2042">https://github.com/w3c/web-platform-tests/commit/8ea6709d2042</a></li>
<li id=l-129><a href=#l-129>#</a> [12:01] -gitbot:#testing- web-platform-tests/noreferrer-name-reuse 8ea6709 Anne van Kesteren: Reusing a window name for a same-origin rel=noreferrer hyperlink</li>
<li id=l-130 class=hide><a href=#l-130>#</a> [12:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-131 class=hide><a href=#l-131>#</a> [12:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-132><a href=#l-132>#</a> [12:02] -gitbot:#testing- [web-platform-tests] annevk opened pull request #2344: Reusing a window name for a same-origin rel=noreferrer hyperlink (master...noreferrer-name-reuse) <a href="https://github.com/w3c/web-platform-tests/pull/2344">https://github.com/w3c/web-platform-tests/pull/2344</a></li>
<li id=l-133 class=hide><a href=#l-133>#</a> [12:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-134 class=hide><a href=#l-134>#</a> [12:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-135><a href=#l-135>#</a> [12:09] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #2345: Update some spec links in html/. (master...spec-html) <a href="https://github.com/w3c/web-platform-tests/pull/2345">https://github.com/w3c/web-platform-tests/pull/2345</a></li>
<li id=l-136 class=hide><a href=#l-136>#</a> [12:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-137 _a="AutomatedTester" _r="zcorpan"><a href=#l-137>#</a> [12:28] &lt;AutomatedTester&gt; zcorpan: hey, I think I finished those elementFromPoint tests</li>
<li id=l-138 _a="AutomatedTester"><a href=#l-138>#</a> [12:28] &lt;AutomatedTester&gt; whenever you get time to review</li>
<li id=l-139 _a="AutomatedTester"><a href=#l-139>#</a> [12:28] &lt;AutomatedTester&gt; even thought I found a bug between browsers but turned out it was a flaky test</li>
<li id=l-140 _a="Ms2ger"><a href=#l-140>#</a> [12:30] &lt;Ms2ger&gt; Aw</li>
<li id=l-141 _a="zcorpan" _r="AutomatedTester"><a href=#l-141>#</a> [12:31] &lt;zcorpan&gt; AutomatedTester: sweet. got a link?</li>
<li id=l-142 _a="AutomatedTester" _r="zcorpan"><a href=#l-142>#</a> [12:31] &lt;AutomatedTester&gt; zcorpan: <a href="https://github.com/w3c/web-platform-tests/pull/2309">https://github.com/w3c/web-platform-tests/pull/2309</a></li>
<li id=l-143 _a="AutomatedTester" _r="Ms2ger"><a href=#l-143>#</a> [12:32] &lt;AutomatedTester&gt; Ms2ger: there might be an inconsistency somewhere but it's not in elementFromPoint</li>
<li id=l-144 _a="AutomatedTester"><a href=#l-144>#</a> [12:32] &lt;AutomatedTester&gt; might be in getBoundingClientRect()</li>
<li id=l-145 _a="zcorpan"><a href=#l-145>#</a> [12:34] &lt;zcorpan&gt; lots of unrelated commits showing up :-(</li>
<li id=l-146 _a="AutomatedTester" _r="zcorpan"><a href=#l-146>#</a> [12:35] &lt;AutomatedTester&gt; zcorpan: yea, I rebased off master and came up</li>
<li id=l-147 _a="AutomatedTester"><a href=#l-147>#</a> [12:35] &lt;AutomatedTester&gt; sorry</li>
<li id=l-148 _a="AutomatedTester"><a href=#l-148>#</a> [12:35] &lt;AutomatedTester&gt; I thought that critic would have noticed it as a rebase</li>
<li id=l-149 _a="AutomatedTester" _r="zcorpan"><a href=#l-149>#</a> [12:36] &lt;AutomatedTester&gt; zcorpan: I can close that PR and recreate it if you would like</li>
<li id=l-150 _a="zcorpan" _r="AutomatedTester"><a href=#l-150>#</a> [12:37] &lt;zcorpan&gt; AutomatedTester: yeah please do that</li>
<li id=l-151 class=hide><a href=#l-151>#</a> [13:04] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-152><a href=#l-152>#</a> [13:04] -gitbot:#testing- [web-platform-tests] AutomatedTester closed pull request #2309: [cssom-view] Add Element From Point tests (master...elementFromPoint) <a href="https://github.com/w3c/web-platform-tests/pull/2309">https://github.com/w3c/web-platform-tests/pull/2309</a></li>
<li id=l-153 class=hide><a href=#l-153>#</a> [13:04] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-154 class=hide><a href=#l-154>#</a> [13:18] * Quits: fwtnb_ (~fwtnb@public.cloak) (&quot;&quot;)</li>
<li id=l-155 class=hide><a href=#l-155>#</a> [13:20] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-156 class=hide><a href=#l-156>#</a> [13:21] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-157 class=hide><a href=#l-157>#</a> [13:31] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-158 class=hide><a href=#l-158>#</a> [14:02] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-159 class=hide><a href=#l-159>#</a> [14:03] * Quits: plh (plehegar@public.cloak) (Client closed connection)</li>
<li id=l-160 class=hide><a href=#l-160>#</a> [14:03] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-161 _a="AutomatedTester"><a href=#l-161>#</a> [14:21] &lt;AutomatedTester&gt; ok... I have put my repo into some weird state</li>
<li id=l-162 _a="AutomatedTester"><a href=#l-162>#</a> [14:21] &lt;AutomatedTester&gt; it says my fork is 10 commits ahead of wpt/master</li>
<li id=l-163 _a="Ms2ger"><a href=#l-163>#</a> [14:22] &lt;Ms2ger&gt; You did do something weird</li>
<li id=l-164 class=hide><a href=#l-164>#</a> [14:23] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-165><a href=#l-165>#</a> [14:23] -gitbot:#testing- [web-platform-tests] AutomatedTester opened pull request #2348: element from point tests (master...dburns/elementFromPoint-tests) <a href="https://github.com/w3c/web-platform-tests/pull/2348">https://github.com/w3c/web-platform-tests/pull/2348</a></li>
<li id=l-166 class=hide><a href=#l-166>#</a> [14:23] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-167 _a="AutomatedTester" _r="zcorpan"><a href=#l-167>#</a> [14:23] &lt;AutomatedTester&gt; zcorpan: all that is of interest is  573fd57 from ^</li>
<li id=l-168 _a="Ms2ger" _r="AutomatedTester"><a href=#l-168>#</a> [14:24] &lt;Ms2ger&gt; AutomatedTester, what does git remote -v say?</li>
<li id=l-169 _a="AutomatedTester" _r="Ms2ger"><a href=#l-169>#</a> [14:24] &lt;AutomatedTester&gt; Ms2ger: the following</li>
<li id=l-170 _a="AutomatedTester"><a href=#l-170>#</a> [14:24] &lt;AutomatedTester&gt; origin	git@github.com:AutomatedTester/web-platform-tests.git (fetch)</li>
<li id=l-171 _a="AutomatedTester"><a href=#l-171>#</a> [14:24] &lt;AutomatedTester&gt; origin	git@github.com:AutomatedTester/web-platform-tests.git (push)</li>
<li id=l-172 _a="AutomatedTester"><a href=#l-172>#</a> [14:24] &lt;AutomatedTester&gt; w3c	git@github.com:w3c/web-platform-tests.git (fetch)</li>
<li id=l-173 _a="AutomatedTester"><a href=#l-173>#</a> [14:24] &lt;AutomatedTester&gt; w3c	git@github.com:w3c/web-platform-tests.git (push)</li>
<li id=l-174 _a="Ms2ger"><a href=#l-174>#</a> [14:24] &lt;Ms2ger&gt; Okay</li>
<li id=l-175 _a="Ms2ger"><a href=#l-175>#</a> [14:24] &lt;Ms2ger&gt; git fetch w3c</li>
<li id=l-176 _a="Ms2ger"><a href=#l-176>#</a> [14:24] &lt;Ms2ger&gt; git checkout dburns/elementFromPoint-tests</li>
<li id=l-177 _a="Ms2ger"><a href=#l-177>#</a> [14:25] &lt;Ms2ger&gt; git rebase -i w3c/master</li>
<li id=l-178 _a="Ms2ger"><a href=#l-178>#</a> [14:25] &lt;Ms2ger&gt; Drop everything except your commit</li>
<li id=l-179 _a="Ms2ger"><a href=#l-179>#</a> [14:25] &lt;Ms2ger&gt; Then git push -f origin dburns/elementFromPoint-tests</li>
<li id=l-180 _a="AutomatedTester" _r="Ms2ger"><a href=#l-180>#</a> [14:25] &lt;AutomatedTester&gt; Ms2ger: that will make critic sad</li>
<li id=l-181 _a="Ms2ger"><a href=#l-181>#</a> [14:26] &lt;Ms2ger&gt; Critic already is sad :)</li>
<li id=l-182 class=hide><a href=#l-182>#</a> [14:31] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-183><a href=#l-183>#</a> [14:31] -gitbot:#testing- [testharness.js] annevk pushed 1 new commit to master: <a href="https://github.com/w3c/testharness.js/commit/8658d6ca31d086e695b333aef55bb388091c7049">https://github.com/w3c/testharness.js/commit/8658d6ca31d086e695b333aef55bb388091c7049</a></li>
<li id=l-184><a href=#l-184>#</a> [14:31] -gitbot:#testing- testharness.js/master 8658d6c Anne van Kesteren: Variable typo in `EventWachter` example</li>
<li id=l-185 _a="zcorpan"><a href=#l-185>#</a> [14:31] &lt;zcorpan&gt; i can review this as is, looking now</li>
<li id=l-186 class=hide><a href=#l-186>#</a> [14:31] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-187 _a="zcorpan" _r="AutomatedTester"><a href=#l-187>#</a> [14:32] &lt;zcorpan&gt; AutomatedTester: ^</li>
<li id=l-188 _a="AutomatedTester"><a href=#l-188>#</a> [14:32] &lt;AutomatedTester&gt; okey dokey</li>
<li id=l-189 _a="AutomatedTester"><a href=#l-189>#</a> [14:32] &lt;AutomatedTester&gt; thanks</li>
<li id=l-190 _a="zcorpan"><a href=#l-190>#</a> [14:36] &lt;zcorpan&gt; is `let` OK in wpt these days?</li>
<li id=l-191 _a="MikeSmith" _r="zcorpan"><a href=#l-191>#</a> [14:36] &lt;MikeSmith&gt; zcorpan: should be, but can't recall that it's come up til now</li>
<li id=l-192 _a="Ms2ger"><a href=#l-192>#</a> [14:37] &lt;Ms2ger&gt; Prefer not to</li>
<li id=l-193 _a="MikeSmith"><a href=#l-193>#</a> [14:37] &lt;MikeSmith&gt; oh</li>
<li id=l-194 _a="MikeSmith"><a href=#l-194>#</a> [14:37] &lt;MikeSmith&gt; why not</li>
<li id=l-195 _a="zcorpan"><a href=#l-195>#</a> [14:37] &lt;zcorpan&gt; doesn't work in IE6</li>
<li id=l-196 _a="zcorpan"><a href=#l-196>#</a> [14:37] * zcorpan ducks</li>
<li id=l-197 _a="Ms2ger"><a href=#l-197>#</a> [14:37] &lt;Ms2ger&gt; Or recent Firefox</li>
<li id=l-198 _a="MikeSmith"><a href=#l-198>#</a> [14:37] &lt;MikeSmith&gt; wha?</li>
<li id=l-199 _a="MikeSmith"><a href=#l-199>#</a> [14:37] &lt;MikeSmith&gt; backed it out?</li>
<li id=l-200 _a="AutomatedTester" _r="Ms2ger"><a href=#l-200>#</a> [14:37] &lt;AutomatedTester&gt; Ms2ger: oh... shu's change hasnt got to release yet?</li>
<li id=l-201 _a="MikeSmith"><a href=#l-201>#</a> [14:37] &lt;MikeSmith&gt; it was there before though</li>
<li id=l-202 _a="AutomatedTester"><a href=#l-202>#</a> [14:38] &lt;AutomatedTester&gt; it's definitely there for nightly</li>
<li id=l-203 _a="AutomatedTester"><a href=#l-203>#</a> [14:38] * AutomatedTester remembers the pain...</li>
<li id=l-204 _a="Ms2ger"><a href=#l-204>#</a> [14:39] &lt;Ms2ger&gt; Has 44 shipped yet?</li>
<li id=l-205 _a="AutomatedTester"><a href=#l-205>#</a> [14:39] &lt;AutomatedTester&gt; no</li>
<li id=l-206 _a="AutomatedTester"><a href=#l-206>#</a> [14:39] &lt;AutomatedTester&gt; that's Aurora</li>
<li id=l-207 _a="Ms2ger"><a href=#l-207>#</a> [14:39] &lt;Ms2ger&gt; <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=932517">https://bugzilla.mozilla.org/show_bug.cgi?id=932517</a> claims let is on by default starting in 44</li>
<li id=l-208 _a="Ms2ger"><a href=#l-208>#</a> [14:39] &lt;Ms2ger&gt; Is that wrong?</li>
<li id=l-209 _a="AutomatedTester"><a href=#l-209>#</a> [14:40] &lt;AutomatedTester&gt; that is right</li>
<li id=l-210 _a="Ms2ger"><a href=#l-210>#</a> [14:40] &lt;Ms2ger&gt; I definitely think tests should work in release Fx, if possible</li>
<li id=l-211 _a="Ms2ger"><a href=#l-211>#</a> [14:40] &lt;Ms2ger&gt; And preferably a few releases back</li>
<li id=l-212 _a="jgraham"><a href=#l-212>#</a> [14:41] &lt;jgraham&gt; Also Safari, which is generally the blocker for fancy ES features</li>
<li id=l-213 _a="AutomatedTester"><a href=#l-213>#</a> [14:41] &lt;AutomatedTester&gt; oh yea... I forgot about Safari</li>
<li id=l-214 _a="MikeSmith"><a href=#l-214>#</a> [14:41] &lt;MikeSmith&gt; hmm me too</li>
<li id=l-215 _a="jgraham"><a href=#l-215>#</a> [14:41] &lt;jgraham&gt; Otherwise we could all write test(() =&gt; {})</li>
<li id=l-216 _a="AutomatedTester" _r="Ms2ger"><a href=#l-216>#</a> [14:41] &lt;AutomatedTester&gt; Ms2ger: why a few releases back?</li>
<li id=l-217 _a="AutomatedTester" _r="zcorpan"><a href=#l-217>#</a> [14:41] &lt;AutomatedTester&gt; zcorpan: I will s/let/var/ then</li>
<li id=l-218 _a="AutomatedTester"><a href=#l-218>#</a> [14:42] &lt;AutomatedTester&gt; fat arrow tests would be nice</li>
<li id=l-219 _a="jgraham"><a href=#l-219>#</a> [14:42] &lt;jgraham&gt; FWIW my unofficial policy has been &quot;latest released version of major browsers&quot;</li>
<li id=l-220 _a="AutomatedTester" _r="Ms2ger"><a href=#l-220>#</a> [14:42] &lt;AutomatedTester&gt; Ms2ger: unless you mean ESR</li>
<li id=l-221 _a="Ms2ger" _r="AutomatedTester"><a href=#l-221>#</a> [14:42] &lt;Ms2ger&gt; AutomatedTester, in case the test points out a regression, it's nice to be able to run it on older releases without touching the test</li>
<li id=l-222 _a="jgraham"><a href=#l-222>#</a> [14:42] * jgraham doesn't care about Fx &lt; stable</li>
<li id=l-223 _a="jgraham" _r="Ms2ger"><a href=#l-223>#</a> [14:43] &lt;jgraham&gt; Ms2ger: For amny of these things, realistically, you will run it on nightly builds with more features preffed on</li>
<li id=l-224 _a="zcorpan" _r="AutomatedTester"><a href=#l-224>#</a> [14:43] &lt;zcorpan&gt; AutomatedTester: reviewed</li>
<li id=l-225 _a="Ms2ger"><a href=#l-225>#</a> [14:44] &lt;Ms2ger&gt; Which is why I don't feel too strongly about &lt; stable</li>
<li id=l-226 _a="Ms2ger"><a href=#l-226>#</a> [14:46] &lt;Ms2ger&gt; I think having tests work a few extra releases is more useful than using `let`, though</li>
<li id=l-227 _a="Ms2ger"><a href=#l-227>#</a> [14:46] &lt;Ms2ger&gt; The tradeoff might be different for other features</li>
<li id=l-228 _a="jgraham" _r="Yeah"><a href=#l-228>#</a> [14:47] &lt;jgraham&gt; Yeah, I agree let is rather low value</li>
<li id=l-229 _a="jgraham"><a href=#l-229>#</a> [14:48] &lt;jgraham&gt; Fat arrow functions seem much more useful as soon as Apple deign to support them</li>
<li id=l-230 _a="AutomatedTester" _r="zcorpan"><a href=#l-230>#</a> [15:00] &lt;AutomatedTester&gt; zcorpan: silly question (mostly because I am stupid), I don't know how to test those other times</li>
<li id=l-231 _a="zcorpan" _r="AutomatedTester"><a href=#l-231>#</a> [15:03] &lt;zcorpan&gt; AutomatedTester: ok hold on</li>
<li id=l-232 _a="AutomatedTester"><a href=#l-232>#</a> [15:03] &lt;AutomatedTester&gt; items*</li>
<li id=l-233 _a="zcorpan"><a href=#l-233>#</a> [15:06] &lt;zcorpan&gt; <a href="http://software.hixie.ch/utilities/js/live-dom-viewer/saved/3751">http://software.hixie.ch/utilities/js/live-dom-viewer/saved/3751</a></li>
<li id=l-234 _a="zcorpan"><a href=#l-234>#</a> [15:09] &lt;zcorpan&gt; <a href="http://software.hixie.ch/utilities/js/live-dom-viewer/saved/3752">http://software.hixie.ch/utilities/js/live-dom-viewer/saved/3752</a></li>
<li id=l-235 _a="AutomatedTester" _r="ahh"><a href=#l-235>#</a> [15:09] &lt;AutomatedTester&gt; ahh, now I get it, ok cool</li>
<li id=l-236 _a="AutomatedTester"><a href=#l-236>#</a> [15:09] &lt;AutomatedTester&gt; thanks zcorpan</li>
<li id=l-237 _a="zcorpan" _r="AutomatedTester"><a href=#l-237>#</a> [15:09] &lt;zcorpan&gt; AutomatedTester: not silly and not stupid, i was just unclear :-)</li>
<li id=l-238 class=hide><a href=#l-238>#</a> [15:24] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-239><a href=#l-239>#</a> [15:24] -gitbot:#testing- [wpt-tools] youennf opened pull request #40: Check clashing test filenames (master...clashing-names-lint) <a href="https://github.com/w3c/wpt-tools/pull/40">https://github.com/w3c/wpt-tools/pull/40</a></li>
<li id=l-240 class=hide><a href=#l-240>#</a> [15:24] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-241 _a="AutomatedTester" _r="zcorpan"><a href=#l-241>#</a> [15:25] &lt;AutomatedTester&gt; zcorpan: should createHTMLDocument() test case return null</li>
<li id=l-242 class=hide><a href=#l-242>#</a> [15:25] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-243><a href=#l-243>#</a> [15:25] -gitbot:#testing- [web-platform-tests] youennf opened pull request #2349: Renaming test filenames that clash when removing extension  (master...fix-clashing-test-filenames) <a href="https://github.com/w3c/web-platform-tests/pull/2349">https://github.com/w3c/web-platform-tests/pull/2349</a></li>
<li id=l-244 class=hide><a href=#l-244>#</a> [15:25] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-245 _a="AutomatedTester"><a href=#l-245>#</a> [15:27] &lt;AutomatedTester&gt; both chrome and Firefox suggest yes</li>
<li id=l-246 _a="AutomatedTester"><a href=#l-246>#</a> [15:27] &lt;AutomatedTester&gt; and safari</li>
<li id=l-247 _a="zcorpan" _r="AutomatedTester"><a href=#l-247>#</a> [15:29] &lt;zcorpan&gt; AutomatedTester: uh, yes, it should return null per spec also.</li>
<li id=l-248 _a="zcorpan"><a href=#l-248>#</a> [15:29] &lt;zcorpan&gt; (step 1)</li>
<li id=l-249 _a="zcorpan"><a href=#l-249>#</a> [15:29] &lt;zcorpan&gt; sorry about that</li>
<li id=l-250 _a="AutomatedTester" _r="zcorpan"><a href=#l-250>#</a> [15:29] &lt;AutomatedTester&gt; zcorpan: your test above has doc.documentElement  === doc.elementFromPoint</li>
<li id=l-251 _a="AutomatedTester"><a href=#l-251>#</a> [15:29] &lt;AutomatedTester&gt; that's why I was checking ;)</li>
<li id=l-252 _a="zcorpan"><a href=#l-252>#</a> [15:29] &lt;zcorpan&gt; yeah my bad :-)</li>
<li id=l-253 _a="AutomatedTester" _r="zcorpan"><a href=#l-253>#</a> [15:30] &lt;AutomatedTester&gt; zcorpan: all done and pushed</li>
<li id=l-254 _a="AutomatedTester"><a href=#l-254>#</a> [15:30] &lt;AutomatedTester&gt; to my pr</li>
<li id=l-255 class=hide><a href=#l-255>#</a> [15:31] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-256 _a="zcorpan" _r="AutomatedTester"><a href=#l-256>#</a> [15:32] &lt;zcorpan&gt; AutomatedTester: new tests look good, but i noticed something else now. the arguments to assert_* are backwards, which will confuse the messages</li>
<li id=l-257 _a="zcorpan"><a href=#l-257>#</a> [15:32] &lt;zcorpan&gt; should be assert_equals(actual, expected, ...)</li>
<li id=l-258 _a="AutomatedTester"><a href=#l-258>#</a> [15:33] &lt;AutomatedTester&gt; hmmm... that is a bizarre order..</li>
<li id=l-259 _a="AutomatedTester"><a href=#l-259>#</a> [15:33] &lt;AutomatedTester&gt; I will reverse them all</li>
<li id=l-260 _a="zcorpan"><a href=#l-260>#</a> [15:34] &lt;zcorpan&gt; LGTM after that. (need to leave now)</li>
<li id=l-261 _a="AutomatedTester" _r="ok"><a href=#l-261>#</a> [15:34] &lt;AutomatedTester&gt; ok, zcorpan I will land then (after this change)</li>
<li id=l-262 _a="zcorpan"><a href=#l-262>#</a> [15:34] &lt;zcorpan&gt; thanks!</li>
<li id=l-263 class=hide><a href=#l-263>#</a> [15:35] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-264 _a="jgraham"><a href=#l-264>#</a> [15:35] &lt;jgraham&gt; For every possible order of arguments to assert functions someone thinks the order is strange and has precedent to back them up</li>
<li id=l-265 _a="AutomatedTester" _r="jgraham"><a href=#l-265>#</a> [15:36] &lt;AutomatedTester&gt; jgraham: I have either worked with them not mattering or expected, actual</li>
<li id=l-266 _a="AutomatedTester"><a href=#l-266>#</a> [15:36] &lt;AutomatedTester&gt; but I agree with your statement</li>
<li id=l-267 _a="jgraham"><a href=#l-267>#</a> [15:36] &lt;jgraham&gt; Well typically it doesn't matter much. It just means that the error message is wrong</li>
<li id=l-268 _a="jgraham"><a href=#l-268>#</a> [15:36] &lt;jgraham&gt; (sometimes it matters more)</li>
<li id=l-269 _a="AutomatedTester" _r="jgraham"><a href=#l-269>#</a> [15:37] &lt;AutomatedTester&gt; jgraham: if I didnt manage the sheriffs I might not care</li>
<li id=l-270 _a="AutomatedTester"><a href=#l-270>#</a> [15:37] &lt;AutomatedTester&gt; about the messages</li>
<li id=l-271 _a="AutomatedTester"><a href=#l-271>#</a> [15:38] &lt;AutomatedTester&gt; but I know what will happen to me</li>
<li id=l-272 _a="AutomatedTester"><a href=#l-272>#</a> [15:48] &lt;AutomatedTester&gt; I do love git and it magically making my commits disappear</li>
<li id=l-273 _a="AutomatedTester"><a href=#l-273>#</a> [15:48] &lt;AutomatedTester&gt; &lt;/sarcasm&gt;</li>
<li id=l-274 class=hide><a href=#l-274>#</a> [16:03] * Quits: MikeSmith (~MikeSmith@public.cloak) (&quot;Lost terminal&quot;)</li>
<li id=l-275 class=hide><a href=#l-275>#</a> [16:04] * Joins: rego (~rego@public.cloak)</li>
<li id=l-276 class=hide><a href=#l-276>#</a> [16:06] * Joins: MikeSmith (~MikeSmith@public.cloak)</li>
<li id=l-277 class=hide><a href=#l-277>#</a> [16:07] * Quits: MikeSmith (~MikeSmith@public.cloak) (&quot;Lost terminal&quot;)</li>
<li id=l-278 class=hide><a href=#l-278>#</a> [16:08] * Joins: MikeSmith (~MikeSmith@public.cloak)</li>
<li id=l-279 class=hide><a href=#l-279>#</a> [16:09] * Quits: MikeSmith (~MikeSmith@public.cloak) (&quot;Lost terminal&quot;)</li>
<li id=l-280 class=hide><a href=#l-280>#</a> [16:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-281><a href=#l-281>#</a> [16:12] -gitbot:#testing- [csswg-test] kojiishi pushed 1 new commit to master: <a href="https://github.com/w3c/csswg-test/commit/bd5659b77e265d8f4ea08b60db0e3207523461fd">https://github.com/w3c/csswg-test/commit/bd5659b77e265d8f4ea08b60db0e3207523461fd</a></li>
<li id=l-282><a href=#l-282>#</a> [16:12] -gitbot:#testing- csswg-test/master bd5659b Koji Ishii: [css-writing-modes-3] Add outline-inline-v(lr|rl)-006...</li>
<li id=l-283 class=hide><a href=#l-283>#</a> [16:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-284 class=hide><a href=#l-284>#</a> [16:14] * Joins: MikeSmith (~MikeSmith@public.cloak)</li>
<li id=l-285 class=hide><a href=#l-285>#</a> [16:17] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-286><a href=#l-286>#</a> [16:17] -gitbot:#testing- [csswg-test] kojiishi pushed 1 new commit to master: <a href="https://github.com/w3c/csswg-test/commit/88f9e57ab438a3f907fe419b9e9b1592618f2047">https://github.com/w3c/csswg-test/commit/88f9e57ab438a3f907fe419b9e9b1592618f2047</a></li>
<li id=l-287><a href=#l-287>#</a> [16:17] -gitbot:#testing- csswg-test/master 88f9e57 Koji Ishii: [css-writing-modes-3] Fix outline-inline-vlr-006...</li>
<li id=l-288 class=hide><a href=#l-288>#</a> [16:18] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-289 class=hide><a href=#l-289>#</a> [16:18] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-290><a href=#l-290>#</a> [16:18] -gitbot:#testing- [web-platform-tests] youennf opened pull request #2350: Fixing link ref=&quot;match&quot; in some reference files (master...fix-ref-info) <a href="https://github.com/w3c/web-platform-tests/pull/2350">https://github.com/w3c/web-platform-tests/pull/2350</a></li>
<li id=l-291 class=hide><a href=#l-291>#</a> [16:18] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-292 class=hide><a href=#l-292>#</a> [16:21] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-293><a href=#l-293>#</a> [16:21] -gitbot:#testing- [csswg-test] kojiishi pushed 1 new commit to master: <a href="https://github.com/w3c/csswg-test/commit/95a081859cc74460a19b194fb8757a53e0648a1b">https://github.com/w3c/csswg-test/commit/95a081859cc74460a19b194fb8757a53e0648a1b</a></li>
<li id=l-294><a href=#l-294>#</a> [16:21] -gitbot:#testing- csswg-test/master 95a0818 Koji Ishii: [css-writing-modes-3] Fixed ref links for outline-inline-v(lr|rl)-006</li>
<li id=l-295 class=hide><a href=#l-295>#</a> [16:21] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-296 class=hide><a href=#l-296>#</a> [16:22] * Quits: rego (~rego@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-297 class=hide><a href=#l-297>#</a> [16:33] * Joins: rego (~rego@public.cloak)</li>
<li id=l-298 class=hide><a href=#l-298>#</a> [16:46] * Joins: ShaneM (~ShaneM@public.cloak)</li>
<li id=l-299 class=hide><a href=#l-299>#</a> [17:03] * Joins: jyasskin_ (~textual@public.cloak)</li>
<li id=l-300 class=hide><a href=#l-300>#</a> [17:03] * Quits: jyasskin_ (~textual@public.cloak) (&quot;Textual IRC Client: www.textualapp.com&quot;)</li>
<li id=l-301 class=hide><a href=#l-301>#</a> [17:42] * Quits: rego (~rego@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-302 class=hide><a href=#l-302>#</a> [18:09] * Joins: ShaneM_ (~ShaneM@public.cloak)</li>
<li id=l-303 class=hide><a href=#l-303>#</a> [18:15] * Quits: ShaneM (~ShaneM@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-304 class=hide><a href=#l-304>#</a> [18:25] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-305 class=hide><a href=#l-305>#</a> [18:29] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;nn&quot;)</li>
<li id=l-306 class=hide><a href=#l-306>#</a> [18:36] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-307 class=hide><a href=#l-307>#</a> [19:04] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-308 class=hide><a href=#l-308>#</a> [19:06] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-309 class=hide><a href=#l-309>#</a> [19:06] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-310 class=hide _a="terri_offline"><a href=#l-310>#</a> [19:10] * terri_offline is now known as terri</li>
<li id=l-311 class=hide _a="terri"><a href=#l-311>#</a> [20:25] * terri is now known as terri_offline</li>
<li id=l-312 class=hide _a="terri_offline"><a href=#l-312>#</a> [20:31] * terri_offline is now known as terri</li>
<li id=l-313 class=hide><a href=#l-313>#</a> [20:33] * Quits: ShaneM_ (~ShaneM@public.cloak) (Client closed connection)</li>
<li id=l-314 class=hide><a href=#l-314>#</a> [20:34] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-315 class=hide><a href=#l-315>#</a> [20:34] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-316 class=hide><a href=#l-316>#</a> [21:00] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-317 class=hide><a href=#l-317>#</a> [21:04] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-318 class=hide><a href=#l-318>#</a> [21:40] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-319 class=hide><a href=#l-319>#</a> [21:41] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-320 class=hide><a href=#l-320>#</a> [21:48] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-321 class=hide><a href=#l-321>#</a> [21:48] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-322><a href=#l-322>#</a> [21:48] -gitbot:#testing- [web-platform-tests] AutomatedTester pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/846c2a11736fd0d9bc9dd09f60a38ab0b4b1dc56">https://github.com/w3c/web-platform-tests/commit/846c2a11736fd0d9bc9dd09f60a38ab0b4b1dc56</a></li>
<li id=l-323><a href=#l-323>#</a> [21:48] -gitbot:#testing- web-platform-tests/master 846c2a1 AutomatedTester: [cssom-view] Add elementFromPoint tests...</li>
<li id=l-324 class=hide><a href=#l-324>#</a> [21:49] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-325 _a="ato" _r="AutomatedTester"><a href=#l-325>#</a> [21:50] &lt;ato&gt; AutomatedTester: ++</li>
<li id=l-326 class=hide><a href=#l-326>#</a> [21:50] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-327><a href=#l-327>#</a> [21:50] -gitbot:#testing- [web-platform-tests] AutomatedTester closed pull request #2348: element from point tests (master...dburns/elementFromPoint-tests) <a href="https://github.com/w3c/web-platform-tests/pull/2348">https://github.com/w3c/web-platform-tests/pull/2348</a></li>
<li id=l-328 class=hide><a href=#l-328>#</a> [21:50] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-329 _a="AutomatedTester" _r="ato"><a href=#l-329>#</a> [21:51] &lt;AutomatedTester&gt; ato: I went to start work on elementsFromPoint and there are no tests for elementFromPoint in m-c</li>
<li id=l-330 class=hide><a href=#l-330>#</a> [21:51] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-331><a href=#l-331>#</a> [21:51] -gitbot:#testing- [web-platform-tests] Wafflespeanut opened pull request #2352: More tests for crypto.getRandomValues() (master...webcrypto) <a href="https://github.com/w3c/web-platform-tests/pull/2352">https://github.com/w3c/web-platform-tests/pull/2352</a></li>
<li id=l-332 class=hide><a href=#l-332>#</a> [21:51] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-333 _a="ato" _r="AutomatedTester"><a href=#l-333>#</a> [21:52] &lt;ato&gt; AutomatedTester: Test-driven development is so overrated, right?</li>
<li id=l-334 _a="ato"><a href=#l-334>#</a> [21:52] &lt;ato&gt; s/Test-driven development/Tests/ I suppose.</li>
<li id=l-335 _a="AutomatedTester" _r="ato"><a href=#l-335>#</a> [21:53] &lt;AutomatedTester&gt; ato: are you suggesting people don't do TDD...</li>
<li id=l-336 _a="AutomatedTester"><a href=#l-336>#</a> [21:55] &lt;AutomatedTester&gt; now to do elementsFromPoint tests</li>
<li id=l-337 class=hide><a href=#l-337>#</a> [22:00] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-338 class=hide><a href=#l-338>#</a> [22:20] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-339 _a="AutomatedTester" _r="zcorpan"><a href=#l-339>#</a> [22:20] &lt;AutomatedTester&gt; zcorpan: those tests have landed</li>
<li id=l-340 _a="zcorpan" _r="AutomatedTester"><a href=#l-340>#</a> [22:20] &lt;zcorpan&gt; AutomatedTester: great, thanks!</li>
<li id=l-341 _a="AutomatedTester" _r="zcorpan"><a href=#l-341>#</a> [22:20] &lt;AutomatedTester&gt; zcorpan: for elementsFromPoint tests, would you prefer a separate file?</li>
<li id=l-342 _a="zcorpan" _r="AutomatedTester"><a href=#l-342>#</a> [22:21] &lt;zcorpan&gt; AutomatedTester: yeah</li>
<li id=l-343 _a="AutomatedTester" _r="zcorpan"><a href=#l-343>#</a> [22:22] &lt;AutomatedTester&gt; zcorpan: since a lot of the tests are copies of the other file can I just CP and then make the changes I need?</li>
<li id=l-344 _a="zcorpan" _r="AutomatedTester"><a href=#l-344>#</a> [22:26] &lt;zcorpan&gt; AutomatedTester: yes. but i can see that there are pros to using the same file, in particular for maintainability. or being DRY in some other way</li>
<li id=l-345 _a="AutomatedTester"><a href=#l-345>#</a> [22:26] &lt;AutomatedTester&gt; this is why I was asking :)</li>
<li id=l-346 _a="zcorpan" _r="yeah"><a href=#l-346>#</a> [22:26] &lt;zcorpan&gt; yeah, it's not always clear for tests which approach is best</li>
<li id=l-347 _a="AutomatedTester"><a href=#l-347>#</a> [22:26] &lt;AutomatedTester&gt; let  me do the tests and then I can come back to you</li>
<li id=l-348 _a="AutomatedTester"><a href=#l-348>#</a> [22:26] &lt;AutomatedTester&gt; since elementsFromPoint will have a more markup probably to get more layers of stuff</li>
<li id=l-349 _a="zcorpan"><a href=#l-349>#</a> [22:26] &lt;zcorpan&gt; making a file very busy is also bad because then it is difficult to debug what is going on</li>
<li id=l-350 _a="AutomatedTester" _r="yea"><a href=#l-350>#</a> [22:27] &lt;AutomatedTester&gt; yea, I will start separate and we can refactor afterwards</li>
<li id=l-351 _a="AutomatedTester"><a href=#l-351>#</a> [22:27] &lt;AutomatedTester&gt; let's get tests and then clean up later</li>
<li id=l-352 _a="zcorpan"><a href=#l-352>#</a> [22:28] &lt;zcorpan&gt; sounds good</li>
<li id=l-353 _a="AutomatedTester"><a href=#l-353>#</a> [22:28] &lt;AutomatedTester&gt; I dont want perfect to be the enemy of good</li>
<li id=l-354 _a="zcorpan"><a href=#l-354>#</a> [22:28] &lt;zcorpan&gt; sure</li>
<li id=l-355 _a="zcorpan"><a href=#l-355>#</a> [22:28] &lt;zcorpan&gt; i know i have copy-pasted tests :-)</li>
<li id=l-356 _a="jgraham"><a href=#l-356>#</a> [22:33] &lt;jgraham&gt; DRY is less valuable for tests</li>
<li id=l-357 _a="jgraham"><a href=#l-357>#</a> [22:34] &lt;jgraham&gt; You generally want them to be immediately understandable rather than composed of many layers of abstraction</li>
<li id=l-358 class=hide><a href=#l-358>#</a> [22:35] * Joins: ShaneM (~ShaneM@public.cloak)</li>
<li id=l-359 _a="AutomatedTester"><a href=#l-359>#</a> [22:37] &lt;AutomatedTester&gt; I copied the tests over, tweaked and have 4 passing tests \o/</li>
<li id=l-360 _a="AutomatedTester"><a href=#l-360>#</a> [22:38] * AutomatedTester mumbles there are 5 failing tests</li>
<li id=l-361 _a="AutomatedTester"><a href=#l-361>#</a> [22:46] &lt;AutomatedTester&gt; all the tests pass</li>
<li id=l-362 _a="AutomatedTester" _r="jgraham"><a href=#l-362>#</a> [22:48] &lt;AutomatedTester&gt; jgraham: how do I disable these tests for m-c?</li>
<li id=l-363 _a="jgraham" _r="AutomatedTester"><a href=#l-363>#</a> [22:51] &lt;jgraham&gt; AutomatedTester: You don't</li>
<li id=l-364 _a="AutomatedTester"><a href=#l-364>#</a> [22:52] &lt;AutomatedTester&gt; awesome</li>
<li id=l-365 _a="AutomatedTester"><a href=#l-365>#</a> [22:52] &lt;AutomatedTester&gt; do I need to do anything to make the cssom tests now run?</li>
<li id=l-366 _a="AutomatedTester"><a href=#l-366>#</a> [22:52] &lt;AutomatedTester&gt; when you do your next merge</li>
<li id=l-367 _a="AutomatedTester"><a href=#l-367>#</a> [22:52] &lt;AutomatedTester&gt; trying to save you some work :)</li>
<li id=l-368 _a="jgraham"><a href=#l-368>#</a> [22:54] &lt;jgraham&gt; No</li>
<li id=l-369 _a="jgraham"><a href=#l-369>#</a> [22:54] &lt;jgraham&gt; Wait</li>
<li id=l-370 _a="jgraham"><a href=#l-370>#</a> [22:54] &lt;jgraham&gt; Are these in wpt or css?</li>
<li id=l-371 _a="AutomatedTester"><a href=#l-371>#</a> [22:54] &lt;AutomatedTester&gt; wpt</li>
<li id=l-372 _a="jgraham" _r="Right"><a href=#l-372>#</a> [22:55] &lt;jgraham&gt; Right, then no.</li>
<li id=l-373 _a="jgraham"><a href=#l-373>#</a> [22:55] &lt;jgraham&gt; For css all bets are off; we will probably never run them</li>
<li id=l-374 _a="AutomatedTester"><a href=#l-374>#</a> [22:55] &lt;AutomatedTester&gt; this is why I put these into WPT</li>
<li id=l-375 _a="jgraham"><a href=#l-375>#</a> [22:55] &lt;jgraham&gt; Or we have no current mechanism to</li>
<li id=l-376 class=hide><a href=#l-376>#</a> [22:55] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-377 _a="jgraham"><a href=#l-377>#</a> [22:55] &lt;jgraham&gt; Obviously I would like us to, and it's not technically impossible</li>
<li id=l-378 _a="AutomatedTester"><a href=#l-378>#</a> [22:58] &lt;AutomatedTester&gt; let</li>
<li id=l-379 _a="AutomatedTester"><a href=#l-379>#</a> [22:58] &lt;AutomatedTester&gt; let's leave that for servo :P</li>
<li id=l-380 class=hide><a href=#l-380>#</a> [23:10] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-381><a href=#l-381>#</a> [23:10] -gitbot:#testing- [web-platform-tests] AutomatedTester opened pull request #2353: [cssom-view] Add elementsFromPoint tests (master...elementsFromPoint) <a href="https://github.com/w3c/web-platform-tests/pull/2353">https://github.com/w3c/web-platform-tests/pull/2353</a></li>
<li id=l-382 class=hide><a href=#l-382>#</a> [23:10] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-383 class=hide><a href=#l-383>#</a> [23:25] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-384 _a="gsnedders"><a href=#l-384>#</a> [23:32] &lt;gsnedders&gt; I do wonder if adding CSS tests to wpt is just asking for a world of pain down the line when we have multiple testsuites for CSS testsuites…</li>
<li id=l-385 class=hide><a href=#l-385>#</a> [23:51] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-386 class=hide><a href=#l-386>#</a> [23:51] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-387 class=hide><a href=#l-387>#</a> [23:58] * Joins: bhill2_ (~bhill2@public.cloak)</li>
<li id=l-388 class=hide><a href=#l-388>#</a> [23:58] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-389 class=hide><a href=#l-389>#</a> Session Close: Wed Nov 18 00:00:00 2015</li>
</ol>
<p id=end><a href=20151116>Previous day</a>, <a href=20151118>Next day</a></p>
<p>Think <a href="/irc-logs/">these logs</a> are useful? Then <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&amp;business=PMXKLAEH53JSS&amp;lc=GB&amp;item_name=Krijn%20Hoetmer&amp;item_number=IRC%20Logs&amp;currency_code=EUR&amp;bn=PP%2dDonationsBF%3abtn_donate_SM%2egif%3aNonHosted"><strong>please donate</strong></a> to show your gratitude (and keep them up, of course). <a href="https://krijnhoetmer.nl/">Thanks! — Krijn</a></p>
<script src=/irc-logs/s2.js></script>
